+++ Scripts/ControllerGameConfigurator.cs
@@ -14,6 +14,7 @@ namespace ControllerSupport
 			Bind<GamepadBuildingPlacementController>().AsSingleton();
 			Bind<GamepadAreaSelectionController>().AsSingleton();
 			Bind<GamepadSelectionController>().AsSingleton();
+			Bind<GamepadValveController>().AsSingleton();
 		}
 	}
 }